fix(mixing): 扩展混搭编码并注册 DLC5 槽位
- 将搜索请求、sidecar 协议和过滤配置中的 `mixing` 改为 `uint64_t`
- 提取 `Setting/MixingCode.hpp` 统一管理 17 个 mixing 槽位与坐标编码范围
- 让校验、世界包络和原生坐标解析按当前支持槽位数处理 DLC5 mixing

+++ src-tauri/src/sidecar.rs
@@ -90,7 +90,7 @@ pub struct SearchRequestPayload {
     pub seed_start: i32,
     pub seed_end: i32,
     #[serde(default)]
-    pub mixing: i32,
+    pub mixing: u64,
     #[serde(default)]
     pub constraints: SearchConstraints,
     #[serde(default)]
@@ -359,7 +359,7 @@ pub struct NormalizedSearchRequestPayload {
     pub world_type: i32,
     pub seed_start: i32,
     pub seed_end: i32,
-    pub mixing: i32,
+    pub mixing: u64,
     #[serde(default)]
     pub required_traits: Vec<String>,
     #[serde(default)]
